refactor: move entity props to each file
