+++ wondrous-app/components/Project/ListMilestone.tsx
@@ -1,0 +1,40 @@
+import Grid from '@mui/material/Grid';
+import TaskCardStatus from 'components/Common/TaskCardStatus';
+import palette from 'theme/palette';
+import { ENTITIES_TYPES, TASK_STATUS_DONE } from 'utils/constants';
+import FlagIcon from 'components/Icons/Sidebar/flag.svg';
+
+import MilestoneProgress from './MilestoneProgress';
+import { useEntityCreateButtonProps, useGetOrgEntity } from './helpers';
+
+const LeftComponent = ({ title }) => (
+  <Grid container gap="12px" alignItems="center" fontWeight="600" color={palette.white}>
+    {title}
+  </Grid>
+);
+
+const RightComponent = ({ type, orgId, status, id }) => (
+  <Grid container>
+    {status === TASK_STATUS_DONE ? (
+      <TaskCardStatus type={type} orgId={orgId} status={status} style={{ background: '#343434' }} />
+    ) : (
+      <MilestoneProgress milestoneId={id} />
+    )}
+  </Grid>
+);
+
+const ListItemMilestone = { LeftComponent, RightComponent };
+
+const useListMilestone = () => ({
+  HeaderTitleProps: {
+    text: 'Milestone',
+    IconComponent: FlagIcon,
+  },
+  CreateButtonProps: useEntityCreateButtonProps(ENTITIES_TYPES.MILESTONE),
+  backgroundImageUrl: '/images/project/milestone-empty-bg.svg',
+  showAllUrl: 'boards?entity=milestone',
+  ListItemComponents: ListItemMilestone,
+  data: useGetOrgEntity('milestone'),
+});
+
+export default useListMilestone;
